+++ lib/lib_display/UDisplay/uDisplay.cpp
@@ -945,7 +945,11 @@ Renderer *uDisplay::Init(void) {
     _panel_config->disp_gpio_num = GPIO_NUM_NC;
 
     _panel_config->flags.disp_active_low = 0;
+#if ESP_IDF_VERSION_MAJOR >= 5
+    _panel_config->flags.refresh_on_demand = 0;
+#else
     _panel_config->flags.relax_on_idle = 0;
+#endif // ESP_IDF_VERSION_MAJOR >= 5
     _panel_config->flags.fb_in_psram = 1;             // allocate frame buffer in PSRAM
 
     ESP_ERROR_CHECK(esp_lcd_new_rgb_panel(_panel_config, &_panel_handle));
@@ -955,9 +959,16 @@ Renderer *uDisplay::Init(void) {
     uint16_t color = random(0xffff);
     ESP_ERROR_CHECK(_panel_handle->draw_bitmap(_panel_handle, 0, 0, 1, 1, &color));
 
+#if ESP_IDF_VERSION_MAJOR < 5
     _rgb_panel = __containerof(_panel_handle, esp_rgb_panel_t, base);
-
     rgb_fb = (uint16_t *)_rgb_panel->fb;
+#else
+    void * buf = NULL;
+    esp_lcd_rgb_panel_get_frame_buffer(_panel_handle, 1, &buf);
+    rgb_fb = (uint16_t *)buf;
+#endif
+
+
 
 #endif // USE_ESP32_S3
   }
@@ -1032,12 +1043,14 @@ Renderer *uDisplay::Init(void) {
 
     esp_lcd_new_i80_bus(&bus_config, &_i80_bus);
 
+#if ESP_IDF_VERSION_MAJOR < 5
     _dma_chan = _i80_bus->dma_chan;
+#endif
 
     uint32_t div_a, div_b, div_n, clkcnt;
     calcClockDiv(&div_a, &div_b, &div_n, &clkcnt, 240*1000*1000, spi_speed*1000000);
     lcd_cam_lcd_clock_reg_t lcd_clock;
-    lcd_clock.lcd_clkcnt_n = std::max(1u, clkcnt - 1);
+    lcd_clock.lcd_clkcnt_n = std::max((uint32_t)1u, clkcnt - 1); // ESP_IDF_VERSION_MAJOR >= 5
     lcd_clock.lcd_clk_equ_sysclk = (clkcnt == 1);
     lcd_clock.lcd_ck_idle_edge = true;
     lcd_clock.lcd_ck_out_edge = false;